v26.05

+++ otherJS/timeAndDate.js
@@ -21,7 +21,7 @@ setInterval(() => {
 
 const now = new Date()
 const options = {
-	weekday: "long", // "long" | "short" | "narrow"
+	weekday: "short", // "long" | "short" | "narrow"
 	month: "short", // "numeric" | "2-digit" | "long" | "short" | "narrow"
 	day: "numeric" // "numeric" | "2-digit"
 	//year: "numeric", // "numeric" | "2-digit"
